a huge commit wwwwww

+++ interpolate_fake_curce.py
@@ -26,7 +26,7 @@ fake_pts_y = []
 print fake_pts_lst_str
 
 for p in fake_pts_lst_str: 
-    that_thing = map(int, p[0:-1].split(", "))
+    that_thing = map(float, p[0:-1].split(", "))
     fake_pts_x.append(that_thing[0])
     fake_pts_y.append(that_thing[1])
